bringup(rednote_hilab_dots.ocr): real_weights attention ok
Tick 45: loaded real layer-0 Qwen2 self-attn weights+QKV biases (params_loaded=5507072); PCC 0.9966 after switching QK^T scores matmul to fp32 (real K +/-320 made bf16 scores the dominant loss). Synthetic regression held 0.99999.

+++ models/demos/rednote_hilab_dots.ocr/tt/attention.py
@@ -238,12 +238,17 @@ class TtAttention(LightweightModule):
         v = self._repeat_kv(v)
 
         # attn = softmax(q k^T * scale + causal_mask) @ v.  k^T -> [1, nh, hd, seq]
+        # Keep the QK^T scores in fp32: real-weight K reaches ±320 (large k_proj
+        # bias), so the pre-softmax scores have a wide dynamic range and bf16
+        # rounding of the scores is the dominant PCC loss (drops ~0.984 -> 0.998
+        # when fp32). fp32_dest_acc already accumulates the dot product in fp32;
+        # storing the matmul output fp32 preserves it through scale/mask/softmax.
         k_t = ttnn.permute(k, (0, 1, 3, 2))
         scores = ttnn.matmul(
             q,
             k_t,
             compute_kernel_config=self.compute_kernel_config,
-            dtype=ttnn.bfloat16,
+            dtype=ttnn.float32,
         )
         scores = ttnn.mul(scores, self.scale)
         scores = ttnn.add(scores, self.attn_mask)
